fastapi-app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-multipart==0.0.6
python-magic-bin==0.4.24; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
celery==5.3.1
redis==4.6.0
requests==2.31.0